Split .net client into source files per class
